added new folder structure and new filenames
the old folder structure was a flat hierarchy inside content/ -- this may lead to a slow system if we have thousands of posts (someday ..). the old filenames were just the timestamp, which was not very human-readable.

the new folder and filename structure now is as follows:

content/{year}/{month}/{year}-{month}-{day}_{hour}-{minute}-{second}_{uniqid}/post.txt

the uniqid is randomly generated and also saved into the textfile for later use

the timestamp and date/time is also saved into the textfile, and this information is no longer read from the filename.

only textfiles with the name 'post.txt' get read to be displayed, so content created before this version is no longer visible