posts: use the slug instead of the id for the permalink and the foldername
previously, we used a unique id for the permalinks and the foldername. now, the post slug is used. if no slug/name is transfered via micropub when creating a new post, we automatically create a slug from the title (if available) or the content text, with a fallback to a uniqid().
if the slug already exists, we append a '-1' (or '-2', or '-3', ...) to the slug.